Race calculations move to separate util file
