+++ publication/src/main/scala/hmda/publication/reports/disclosure/D51.scala
@@ -10,6 +10,7 @@ import hmda.model.publication.reports.ReportTypeEnum.Disclosure
 import hmda.model.publication.reports._
 import hmda.publication.reports.util.DateUtil._
 import hmda.publication.reports.util.DispositionType._
+import hmda.publication.reports.util.RaceUtil._
 import hmda.publication.reports.util.ReportUtil._
 import hmda.query.model.filing.LoanApplicationRegisterQuery
